requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
python-multipart
pydantic
python-dotenv
tqdm
numpy
